1.add direct validations whith contacts.js which has two diferent joischemes; 2. add direct middleware whith file validation.js(middleware, which get joiScheme and doing validation)
