+++ middlewares/index.js
@@ -1,0 +1,3 @@
+const validation = require("./validation");
+
+module.exports = { validation };
