+++ validations/index.js
@@ -1,0 +1,3 @@
+const validationsSchemes = require("./contacts");
+
+module.exports = { validationsSchemes };
